Trait hir::db::ExpandDatabase

pub trait ExpandDatabase: Database + HasQueryGroup<ExpandDatabaseStorage> + SourceDatabase {

    // Required methods
    fn ast_id_map(&self, file_id: HirFileId) -> Arc<AstIdMap>;
    fn parse_or_expand(&self, file_id: HirFileId) -> SyntaxNode<RustLanguage>;
    fn parse_or_expand_with_err(
        &self,
        file_id: HirFileId
    ) -> ValueResult<Parse<SyntaxNode<RustLanguage>>, ExpandError>;
    fn parse_macro_expansion(
        &self,
        macro_file: MacroFile
    ) -> ValueResult<(Parse<SyntaxNode<RustLanguage>>, Arc<TokenMap>), ExpandError>;
    fn intern_macro_call(&self, macro_call: MacroCallLoc) -> MacroCallId;
    fn lookup_intern_macro_call(&self, key: MacroCallId) -> MacroCallLoc;
    fn macro_arg(
        &self,
        id: MacroCallId
    ) -> ValueResult<Option<Arc<(Subtree<TokenId>, TokenMap, SyntaxFixupUndoInfo)>>, Arc<Box<[SyntaxError], Global>>>;
    fn macro_arg_node(
        &self,
        id: MacroCallId
    ) -> ValueResult<Option<GreenNode>, Arc<Box<[SyntaxError], Global>>>;
    fn macro_expander(&self, id: MacroDefId) -> TokenExpander;
    fn decl_macro_expander(
        &self,
        def_crate: Idx<CrateData>,
        id: InFile<FileAstId<Macro>>
    ) -> Arc<DeclarativeMacroExpander>;
    fn macro_expand(
        &self,
        macro_call: MacroCallId
    ) -> ValueResult<Arc<Subtree<TokenId>>, ExpandError>;
    fn include_expand(
        &self,
        arg_id: MacroCallId
    ) -> Result<(Arc<(Subtree<TokenId>, TokenMap)>, FileId), ExpandError>;
    fn expand_proc_macro(
        &self,
        call: MacroCallId
    ) -> ValueResult<Arc<Subtree<TokenId>>, ExpandError>;
    fn parse_macro_expansion_error(
        &self,
        macro_call: MacroCallId
    ) -> ValueResult<Box<[SyntaxError], Global>, ExpandError>;
    fn hygiene_frame(&self, file_id: HirFileId) -> Arc<HygieneFrame>;

}

Required Methods

fn ast_id_map(&self, file_id: HirFileId) -> Arc<AstIdMap>

fn parse_or_expand(&self, file_id: HirFileId) -> SyntaxNode<RustLanguage>

Main public API – parses a hir file, not caring whether it’s a real file or a macro expansion.

fn parse_or_expand_with_err( &self, file_id: HirFileId ) -> ValueResult<Parse<SyntaxNode<RustLanguage>>, ExpandError>

fn parse_macro_expansion( &self, macro_file: MacroFile ) -> ValueResult<(Parse<SyntaxNode<RustLanguage>>, Arc<TokenMap>), ExpandError>

Implementation for the macro case.

fn intern_macro_call(&self, macro_call: MacroCallLoc) -> MacroCallId

Macro ids. That’s probably the tricksiest bit in rust-analyzer, and the reason why we use salsa at all.

We encode macro definitions into ids of macro calls, this what allows us to be incremental.

fn lookup_intern_macro_call(&self, key: MacroCallId) -> MacroCallLoc

fn macro_arg( &self, id: MacroCallId ) -> ValueResult<Option<Arc<(Subtree<TokenId>, TokenMap, SyntaxFixupUndoInfo)>>, Arc<Box<[SyntaxError], Global>>>

Lowers syntactic macro call to a token tree representation.

fn macro_arg_node( &self, id: MacroCallId ) -> ValueResult<Option<GreenNode>, Arc<Box<[SyntaxError], Global>>>

Extracts syntax node, corresponding to a macro call. That’s a firewall query, only typing in the macro call itself changes the returned subtree.

fn macro_expander(&self, id: MacroDefId) -> TokenExpander

Fetches the expander for this macro.

fn decl_macro_expander( &self, def_crate: Idx<CrateData>, id: InFile<FileAstId<Macro>> ) -> Arc<DeclarativeMacroExpander>

Fetches (and compiles) the expander of this decl macro.

fn macro_expand( &self, macro_call: MacroCallId ) -> ValueResult<Arc<Subtree<TokenId>>, ExpandError>

Expand macro call to a token tree.

fn include_expand( &self, arg_id: MacroCallId ) -> Result<(Arc<(Subtree<TokenId>, TokenMap)>, FileId), ExpandError>

fn expand_proc_macro( &self, call: MacroCallId ) -> ValueResult<Arc<Subtree<TokenId>>, ExpandError>

Special case of the previous query for procedural macros. We can’t LRU proc macros, since they are not deterministic in general, and non-determinism breaks salsa in a very, very, very bad way. @edwin0cheng heroically debugged this once! See #4315 for details

fn parse_macro_expansion_error( &self, macro_call: MacroCallId ) -> ValueResult<Box<[SyntaxError], Global>, ExpandError>

Firewall query that returns the errors from the parse_macro_expansion query.

fn hygiene_frame(&self, file_id: HirFileId) -> Arc<HygieneFrame>