Enum hir::Adt

pub enum Adt {
    Struct(Struct),
    Union(Union),
    Enum(Enum),
}

A Data Type

Variants

Struct(Struct)

Union(Union)

Enum(Enum)

Implementations

impl Adt

pub fn has_non_default_type_params(self, db: &dyn HirDatabase) -> bool

pub fn ty(self, db: &dyn HirDatabase) -> Type

Turns this ADT into a type. Any type parameters of the ADT will be turned into unknown types, which is good for e.g. finding the most general set of completions, but will not look very nice when printed.

pub fn module(self, db: &dyn HirDatabase) -> Module

pub fn krate(self, db: &dyn HirDatabase) -> Crate

pub fn name(self, db: &dyn HirDatabase) -> Name