hir

Enum MirEvalError

pub enum MirEvalError {

    ConstEvalError(String, Box<ConstEvalError>),
    LayoutError(LayoutError, StoredTy),
    TargetDataLayoutNotAvailable(TargetLoadError),
    UndefinedBehavior(String),
    Panic(String),
    MirLowerError(FunctionId, MirLowerError),
    MirLowerErrorForClosure(InternedClosureId, MirLowerError),
    TypeIsUnsized(StoredTy, &'static str),
    NotSupported(String),
    InvalidConst(StoredConst),
    InFunction(Box<MirEvalError>, Vec<(Either<FunctionId, InternedClosureId>, MirSpan, DefWithBodyId)>),
    ExecutionLimitExceeded,
    StackOverflow,
    InvalidVTableId(usize),
    CoerceUnsizedError(StoredTy),
    InternalError(Box<str>),

}

Variants

ConstEvalError(String, Box<ConstEvalError>)

LayoutError(LayoutError, StoredTy)

TargetDataLayoutNotAvailable(TargetLoadError)

UndefinedBehavior(String)

Means that code had undefined behavior. We donโt try to actively detect UB, but if it was detected then use this type of error.

Panic(String)

MirLowerError(FunctionId, MirLowerError)

MirLowerErrorForClosure(InternedClosureId, MirLowerError)

TypeIsUnsized(StoredTy, &'static str)

NotSupported(String)

InvalidConst(StoredConst)

InFunction(Box<MirEvalError>, Vec<(Either<FunctionId, InternedClosureId>, MirSpan, DefWithBodyId)>)

ExecutionLimitExceeded

StackOverflow

InvalidVTableId(usize)

FIXME: Fold this into InternalError

CoerceUnsizedError(StoredTy)

?

InternalError(Box<str>)

These should not occur, usually indicates a bug in mir lowering.

Implementations

impl MirEvalError

pub fn pretty_print( &self, f: &mut String, db: &(dyn HirDatabase + 'static), span_formatter: impl Fn(FileId, TextRange) -> String, display_target: DisplayTarget, ) -> Result<(), Error>

pub fn is_panic(&self) -> Option<&str>