Enum hir::TypeRef

pub enum TypeRef {

    Never,
    Placeholder,
    Tuple(Vec<TypeRef, Global>),
    Path(Path),
    RawPtr(Box<TypeRef, Global>, Mutability),
    Reference(Box<TypeRef, Global>, Option<LifetimeRef>, Mutability),
    Array(Box<TypeRef, Global>, ConstScalar),
    Slice(Box<TypeRef, Global>),
    Fn(Vec<TypeRef, Global>, bool),
    ImplTrait(Vec<Interned<TypeBound>, Global>),
    DynTrait(Vec<Interned<TypeBound>, Global>),
    Macro(InFile<FileAstId<MacroCall>>),
    Error,

}

Compare ty::Ty

Note: Most users of TypeRef that end up in the salsa database intern it using Interned<TypeRef> to save space. But notably, nested TypeRefs are not interned, since that does not seem to save any noticeable amount of memory.

Variants

Never

Placeholder

Tuple(Vec<TypeRef, Global>)

Path(Path)

RawPtr(Box<TypeRef, Global>, Mutability)

Reference(Box<TypeRef, Global>, Option<LifetimeRef>, Mutability)

Array(Box<TypeRef, Global>, ConstScalar)

Slice(Box<TypeRef, Global>)

Fn(Vec<TypeRef, Global>, bool)

A fn pointer. Last element of the vector is the return type.

ImplTrait(Vec<Interned<TypeBound>, Global>)

DynTrait(Vec<Interned<TypeBound>, Global>)

Macro(InFile<FileAstId<MacroCall>>)

Error

Implementations

impl TypeRef

pub fn from_ast(ctx: &LowerCtx<'_>, node: Type) -> TypeRef

Converts an ast::TypeRef to a hir::TypeRef.

pub fn walk(&self, f: &mut impl FnMut(&TypeRef))