hir::import_map
Function search_dependencies

pub fn search_dependencies(
    db: &dyn DefDatabase,
    krate: Idx<CrateData>,
    query: &Query,
) -> HashSet<ItemInNs, FxBuildHasher>

Searches dependencies of krate for an importable name matching query.

This returns a list of items that could be imported from dependencies of krate.