hir

Module next_solver

Things relevant to the next trait solver.

Modules

abi

ABI-related things in the next-trait-solver.

fold

Fold impls for the next-trait-solver.

format_proof_tree

fulfill

Fulfill loop for next-solver.

generics

Things related to generics in the next-trait-solver.

infer

Infer context the next-trait-solver.

interner

Things related to the Interner in the next-trait-solver.

normalize

obligation_ctxt

predicate

Things related to predicates.

util

Various utilities for the next-trait-solver.

Structs

AdtDef

AdtDefInner

AdtFlags

AdtIdWrapper

AllocId

BoundConst

BoundExistentialPredicates

BoundExistentialPredicatesStorage

BoundRegion

BoundTy

BoundVarKinds

BoundVarKindsStorage

CallableIdWrapper

CanonicalVars

CanonicalVarsStorage

Clause

Clauses

ClausesCachedTypeInfo

ClausesStorage

ClosureIdWrapper

Const

ConstBytes

CoroutineIdWrapper

DbInterner

DefaultAny

DefaultConsts

DefaultEmpty

DefaultRegions

DefaultTypes

DepNodeIndex

EarlyParamRegion

ErrorGuaranteed

ExprConst

ExternalConstraints

Features

GeneralConstIdWrapper

GenericArg

GenericArgs

GenericArgsStorage

LateParamRegion

The parameter representation of late-bound function parameters, “some region at least as big as the scope fr.scope”.

ParamConst

ParamEnv

ParamEnvAnd

ParamTy

PatList

PatListStorage

Pattern

Placeholder

PredefinedOpaques

PredefinedOpaquesStorage

Predicate

Region

RegionAssumptions

RegionAssumptionsStorage

SolverDefIds

SolverDefIdsStorage

Span

StoredBoundExistentialPredicates

StoredBoundVarKinds

StoredCanonicalVars

StoredClauses

StoredConst

StoredEarlyBinder

StoredGenericArg

StoredGenericArgs

StoredPatList

StoredPolyFnSig

StoredPredefinedOpaques

StoredRegion

StoredRegionAssumptions

StoredSolverDefIds

StoredTy

StoredTys

StoredVariancesOf

Term

Tracked

TraitIdWrapper

Ty

TypeAliasIdWrapper

Tys

TysStorage

UnsizingParams

Valtree

ValueConst

A type-level constant value.

VariancesOf

VariancesOfStorage

VariantIdx

Enums

AnyImplId

BoundRegionKind

BoundTyKind

BoundVarKind

Ctor

Mutability

SolverDefId

TermVid

VariantDef

Traits

WorldExposer

This is a visitor trait that treats any interned thing specifically. Visitables are expected to call the trait’s methods when encountering an interned. This is used to implement marking in GC.

Functions

attach_db

attach_db_allow_change

clear_tls_solver_cache

Clears the thread-local trait solver cache.

collect_ty_garbage⚠

Safety

default_types

error_for_param_kind

mk_param

references_non_lt_error

with_attached_db

Type Aliases

AliasTerm

AliasTy

ArgOutlivesPredicate

Binder

BoundExistentialPredicate

Canonical

CanonicalQueryInput

CanonicalVarKind

CanonicalVarValues

ClauseKind

CoercePredicate

ConstKind

EarlyBinder

ExistentialPredicate

ExistentialProjection

ExistentialTraitRef

ExternalConstraintsData

FnHeader

FnSig

FxIndexMap

GenericArgKind

Goal

ImplTraitIdx

NormalizesTo

OpaqueTypeKey

OutlivesPredicate

PatternKind

PlaceholderConst

PlaceholderRegion

PlaceholderTy

PolyCoercePredicate

PolyExistentialProjection

PolyExistentialTraitRef

PolyFnSig

PolyProjectionPredicate

PolyRegionOutlivesPredicate

PolySubtypePredicate

PolyTraitPredicate

PolyTraitRef

PolyTypeOutlivesPredicate

PredicateKind

ProjectionPredicate

QueryResult

RegionKind

RegionOutlivesPredicate

SimplifiedType

SubtypePredicate

TermKind

TraitPredicate

TraitRef

TyKind

TypeError

TypeOutlivesPredicate

TypingMode

UnevaluatedConst