hir::next_solver

Module infer

Infer context the next-trait-solver.

Modules

at

A nice interface for working with the infcx. The basic idea is to do infcx.at(cause, param_env), which sets the “cause” of the operation as well as the surrounding parameter environment. Then you can do something like .sub(a, b) or .eq(a, b) to create a subtype or equality relationship respectively. The first argument is always the “expected” output from the POV of diagnostics.

canonical

Canonicalization is the key to constructing a query in the middle of type inference. Ordinarily, it is not possible to store types from type inference in query keys, because they contain references to inference variables whose lifetimes are too short and so forth. Canonicalizing a value T1 using canonicalize_query produces two things:

opaque_types

Things related to the infer context of the next-trait-solver.

region_constraints

See README.md.

relate

This module contains the definitions of most TypeRelations in the type system (except for some relations used for diagnostics and heuristics in the compiler). As well as the implementation of Relate for interned things (Ty/Const/etc).

resolve

Things for resolving vars in the infer context of the next-trait-solver.

Structs

FixupError

InferCtxt

InferCtxtBuilder

Used to configure inference contexts before their creation.

InferCtxtInner

This type contains all the things within InferCtxt that sit within a RefCell and are involved with taking/rolling back snapshots. Snapshot operations are hot enough that we want only one call to borrow_mut per call to start_snapshot and rollback_to.

InferOk

InferOk<'db, ()> is used a lot. It may seem like a useless wrapper around PredicateObligations, but it has one important property: because InferOk is marked with #[must_use], if you have a method InferCtxt::f that returns InferResult<()> and you call it with infcx.f()?; you’ll get a warning about the obligations being discarded without use, which is probably unintentional and has been a source of bugs in the past.

MemberConstraint

Requires that region must be equal to one of the regions in choice_regions. We often denote this using the syntax:

TypeOutlivesConstraint

See the region_obligations field for more information.

TypeTrace

The trace designates the path through inference that we took to encounter an error or subtyping constraint.

Enums

BoundRegionConversionTime

Times when we replace bound regions with existentials:

TyOrConstInferVar

Helper for InferCtxt::ty_or_const_infer_var_changed (see comment on that), currently used only for traits::fulfill’s list of stalled_on inference variables.

ValuePairs

See the error_reporting module for more details.

Traits

DbInternerInferExt

Type Aliases

InferResult

ObligationInspector

A callback that can be provided to inspect_typeck. Invoked on evaluation of root obligations.