hir::next_solver::infer::canonical

Module canonicalizer

This module contains code to canonicalize values into a Canonical<'db, T>.

For an overview of what canonicalization is and how it fits into rustc, check out the chapter in the rustc dev guide.

Structs

OriginalQueryValues

When we canonicalize a value to form a query, we wind up replacing various parts of it with canonical variables. This struct stores those replaced bits to remember for when we process the query result.