hir::next_solver::infer

Module traits

Trait Resolution. See the rustc-dev-guide for more information on how this works.

Structs

Obligation

An Obligation represents some trait reference (e.g., i32: Eq) for which the “impl_source” must be found. The process of finding an “impl_source” is called “resolving” the Obligation. This process consists of either identifying an impl (e.g., impl Eq for i32) that satisfies the obligation, or else finding a bound that is in scope. The eventual result is usually a Selection (defined below).

ObligationCause

The reason why we incurred this obligation; used for error reporting.

Type Aliases

ObligationInspector

A callback that can be provided to inspect_typeck. Invoked on evaluation of root obligations.