hir::next_solver::normalize

Function deeply_normalize_with_skipped_universes

pub fn deeply_normalize_with_skipped_universes<'db, T>(
    at: At<'_, 'db>,
    value: T,
    universes: Vec<Option<UniverseIndex>>,
) -> Result<T, Vec<NextSolverError<'db>>>
where
    T: TypeFoldable<DbInterner<'db>>,

Deeply normalize all aliases in value. This does not handle inference and expects its input to be already fully resolved.

Additionally takes a list of universes which represents the binders which have been entered before passing value to the function. This is currently needed for normalize_erasing_regions, which skips binders as it walks through a type.