hir::next_solver::predicate

Type Alias PredicateKind

pub type PredicateKind<'db> = PredicateKind<DbInterner<'db>>;

Aliased Type

pub enum PredicateKind<'db> {
    Clause(ClauseKind<DbInterner<'db>>),
    DynCompatible(TraitIdWrapper),
    Subtype(SubtypePredicate<DbInterner<'db>>),
    Coerce(CoercePredicate<DbInterner<'db>>),
    ConstEquate(Const<'db>, Const<'db>),
    Ambiguous,
    NormalizesTo(NormalizesTo<DbInterner<'db>>),
    AliasRelate(Term<'db>, Term<'db>, AliasRelationDirection),
}

Variants

Clause(ClauseKind<DbInterner<'db>>)

Prove a clause

DynCompatible(TraitIdWrapper)

Trait must be dyn-compatible.

Subtype(SubtypePredicate<DbInterner<'db>>)

T1 <: T2

This obligation is created most often when we have two unresolved type variables and hence don’t have enough information to process the subtyping obligation yet.

Coerce(CoercePredicate<DbInterner<'db>>)

T1 coerced to T2

Like a subtyping obligation, this is created most often when we have two unresolved type variables and hence don’t have enough information to process the coercion obligation yet. At the moment, we actually process coercions very much like subtyping and don’t handle the full coercion logic.

ConstEquate(Const<'db>, Const<'db>)

Constants must be equal. The first component is the const that is expected.

Ambiguous

A marker predicate that is always ambiguous. Used for coherence to mark opaque types as possibly equal to each other but ambiguous.

NormalizesTo(NormalizesTo<DbInterner<'db>>)

This should only be used inside of the new solver for AliasRelate and expects the term to be always be an unconstrained inference variable. It is used to normalize alias as much as possible. In case the alias is rigid - i.e. it cannot be normalized in the current environment - this constrains term to be equal to the alias itself.

It is likely more useful to think of this as a function normalizes_to(alias), whose return value is written into term.

AliasRelate(Term<'db>, Term<'db>, AliasRelationDirection)

Separate from ClauseKind::Projection which is used for normalization in new solver. This predicate requires two terms to be equal to eachother.

Only used for new solver.