hir::next_solver

Struct DbInterner

pub struct DbInterner<'db> { /* private fields */ }

Implementations

impl<'db> DbInterner<'db>

pub fn instantiate_bound_regions<T, F>( self, value: Binder<DbInterner<'db>, T>, fld_r: F, ) -> (T, IndexMap<BoundRegion, Region<'db>, BuildHasherDefault<FxHasher>>)
where F: FnMut(BoundRegion) -> Region<'db>, T: TypeFoldable<DbInterner<'db>>,

Replaces all regions bound by the given Binder with the results returned by the closure; the closure is expected to return a free region (relative to this binder), and hence the binder is removed in the return type. The closure is invoked once for each unique BoundRegionKind; multiple references to the same BoundRegionKind will reuse the previous result. A map is returned at the end with each bound region and the free region that replaced it.

Panics

This method only replaces late bound regions. Any types or constants bound by value will cause an ICE.

pub fn instantiate_bound_regions_uncached<T, F>( self, value: Binder<DbInterner<'db>, T>, replace_regions: F, ) -> T
where F: FnMut(BoundRegion) -> Region<'db>, T: TypeFoldable<DbInterner<'db>>,

pub fn instantiate_bound_regions_with_erased<T>( self, value: Binder<DbInterner<'db>, T>, ) -> T
where T: TypeFoldable<DbInterner<'db>>,

Replaces any late-bound regions bound in value with 'erased. Useful in codegen but also method lookup and a few other places where precise region relationships are not required.

impl<'db> DbInterner<'db>

pub fn conjure() -> DbInterner<'db>

pub fn new_with( db: &'db (dyn HirDatabase + 'static), krate: Option<Crate>, block: Option<BlockId>, ) -> DbInterner<'db>

pub fn db(&self) -> &'db (dyn HirDatabase + 'static)

impl<'db> DbInterner<'db>

pub fn shift_bound_var_indices<T>(self, bound_vars: usize, value: T) -> T
where T: TypeFoldable<DbInterner<'db>>,

pub fn replace_escaping_bound_vars_uncached<T>( self, value: T, delegate: impl BoundVarReplacerDelegate<'db>, ) -> T
where T: TypeFoldable<DbInterner<'db>>,

pub fn replace_bound_vars_uncached<T>( self, value: Binder<DbInterner<'db>, T>, delegate: impl BoundVarReplacerDelegate<'db>, ) -> T
where T: TypeFoldable<DbInterner<'db>>,

pub fn mk_fn_sig<I>( self, inputs: I, output: Ty<'db>, c_variadic: bool, safety: Safety, abi: FnAbi, ) -> FnSig<DbInterner<'db>>
where I: IntoIterator<Item = Ty<'db>>,