hir::next_solver

Type Alias BoundRegionKind

pub type BoundRegionKind<'db> = BoundRegionKind<DbInterner<'db>>;

Aliased Type

pub enum BoundRegionKind<'db> {
    Anon,
    NamedForPrinting(Symbol),
    Named(SolverDefId),
    ClosureEnv,
}

Variants

Anon

An anonymous region parameter for a given fn (&T)

NamedForPrinting(Symbol)

An anonymous region parameter with a Symbol name.

Used to give late-bound regions names for things like pretty printing.

Named(SolverDefId)

Late-bound regions that appear in the AST.

ClosureEnv

Anonymous region for the implicit env pointer parameter to a closure