hir::next_solver

Type Alias CanonicalVarKind

pub type CanonicalVarKind<'db> = CanonicalVarKind<DbInterner<'db>>;

Aliased Type

pub enum CanonicalVarKind<'db> {
    Ty {
        ui: UniverseIndex,
        sub_root: BoundVar,
    },
    Int,
    Float,
    PlaceholderTy(Placeholder<DbInterner<'db>, BoundTy<DbInterner<'db>>>),
    Region(UniverseIndex),
    PlaceholderRegion(Placeholder<DbInterner<'db>, BoundRegion<DbInterner<'db>>>),
    Const(UniverseIndex),
    PlaceholderConst(Placeholder<DbInterner<'db>, BoundConst<DbInterner<'db>>>),
}

Variants

Ty

General type variable ?T that can be unified with arbitrary types.

We also store the index of the first type variable which is sub-unified with this one. If there is no inference variable related to this one, its sub_root just points to itself.

Fields

ui: UniverseIndex

sub_root: BoundVar

Int

Integral type variable ?I (that can only be unified with integral types).

Float

Floating-point type variable ?F (that can only be unified with float types).

PlaceholderTy(Placeholder<DbInterner<'db>, BoundTy<DbInterner<'db>>>)

A “placeholder” that represents “any type”.

Region(UniverseIndex)

Region variable '?R.

PlaceholderRegion(Placeholder<DbInterner<'db>, BoundRegion<DbInterner<'db>>>)

A “placeholder” that represents “any region”. Created when you are solving a goal like for<'a> T: Foo<'a> to represent the bound region 'a.

Const(UniverseIndex)

Some kind of const inference variable.

PlaceholderConst(Placeholder<DbInterner<'db>, BoundConst<DbInterner<'db>>>)

A “placeholder” that represents “any const”.