hir

Module semantics

See Semantics.

Modules

child_by_source 🔒

When constructing hir, we start at some parent syntax node and recursively lower the children.

source_to_def 🔒

Maps syntax of various definitions to their semantic ids.

Macros

to_def_impls 🔒

Structs

PathResolutionPerNs

RenameConflictsVisitor 🔒

Semantics

Primary API to get semantic information, like types, from syntax trees.

SemanticsImpl

SemanticsScope

SemanticsScope encapsulates the notion of a scope (the set of visible names) at a particular program point.

TypeInfo

VisibleTraits

Enums

LintAttr

PathResolution

Constants

CONTINUE_NO_BREAKS 🔒

Traits

ToDef

Functions

find_root 🔒

macro_call_to_macro_id 🔒