Struct hir::CfgOptions

pub struct CfgOptions { /* private fields */ }

Configuration options used for conditional compilation on items with cfg attributes. We have two kind of options in different namespaces: atomic options like unix, and key-value options like target_arch="x86".

Note that for key-value options, one key can have multiple values (but not none). feature is an example. We have both feature="foo" and feature="bar" if features foo and bar are both enabled. And here, we store key-value options as a set of tuple of key and value in key_values.

See: https://doc.rust-lang.org/reference/conditional-compilation.html#set-configuration-options

Implementations

impl CfgOptions

pub fn check(&self, cfg: &CfgExpr) -> Option<bool>

pub fn insert_atom(&mut self, key: SmolStr)

pub fn insert_key_value(&mut self, key: SmolStr, value: SmolStr)

pub fn apply_diff(&mut self, diff: CfgDiff)

pub fn get_cfg_keys(&self) -> impl Iterator<Item = &SmolStr>

pub fn get_cfg_values<'a>( &'a self, cfg_key: &'a str ) -> impl Iterator<Item = &'a SmolStr> + 'a