Struct hir::Function

pub struct Function { /* private fields */ }

Implementations

impl Function

pub fn module(self, db: &dyn HirDatabase) -> Module

pub fn name(self, db: &dyn HirDatabase) -> Name

pub fn ty(self, db: &dyn HirDatabase) -> Type

pub fn fn_ptr_type(self, db: &dyn HirDatabase) -> Type

pub fn ret_type(self, db: &dyn HirDatabase) -> Type

Get this function’s return type

pub fn ret_type_with_args( self, db: &dyn HirDatabase, generics: impl Iterator<Item = Type>, ) -> Type

pub fn async_ret_type(self, db: &dyn HirDatabase) -> Option<Type>

pub fn has_self_param(self, db: &dyn HirDatabase) -> bool

pub fn self_param(self, db: &dyn HirDatabase) -> Option<SelfParam>

pub fn assoc_fn_params(self, db: &dyn HirDatabase) -> Vec<Param>

pub fn num_params(self, db: &dyn HirDatabase) -> usize

pub fn method_params(self, db: &dyn HirDatabase) -> Option<Vec<Param>>

pub fn params_without_self(self, db: &dyn HirDatabase) -> Vec<Param>

pub fn params_without_self_with_args( self, db: &dyn HirDatabase, generics: impl Iterator<Item = Type>, ) -> Vec<Param>

pub fn is_const(self, db: &dyn HirDatabase) -> bool

pub fn is_async(self, db: &dyn HirDatabase) -> bool

pub fn returns_impl_future(self, db: &dyn HirDatabase) -> bool

pub fn is_test(self, db: &dyn HirDatabase) -> bool

Does this function have #[test] attribute?

pub fn is_main(self, db: &dyn HirDatabase) -> bool

is this a fn main or a function with an export_name of main?

pub fn exported_main(self, db: &dyn HirDatabase) -> bool

Is this a function with an export_name of main?

pub fn is_ignore(self, db: &dyn HirDatabase) -> bool

Does this function have the ignore attribute?

pub fn is_bench(self, db: &dyn HirDatabase) -> bool

Does this function have #[bench] attribute?

pub fn is_unstable(self, db: &dyn HirDatabase) -> bool

Is this function marked as unstable with #[feature] attribute?

pub fn is_unsafe_to_call(self, db: &dyn HirDatabase) -> bool

pub fn has_body(self, db: &dyn HirDatabase) -> bool

Whether this function declaration has a definition.

This is false in the case of required (not provided) trait methods.

pub fn as_proc_macro(self, db: &dyn HirDatabase) -> Option<Macro>

pub fn eval( self, db: &dyn HirDatabase, span_formatter: impl Fn(FileId, TextRange) -> String, ) -> String