hir

Struct InFileWrapper

pub struct InFileWrapper<FileKind, T> {
    pub file_id: FileKind,
    pub value: T,
}

InFile<T> stores a value of T inside a particular file/syntax tree.

Typical usages are:

InFile<SyntaxNode> – syntax node in a file

InFile<ast::FnDef> – ast node in a file

InFile<TextSize> – offset in a file

Fields
file_id: FileKindvalue: T
Implementations

impl<N> InFileWrapper<HirFileId, FileAstId<N>>
where N: AstNode,

pub fn to_node(&self, db: &(dyn ExpandDatabase + 'static)) -> N

pub fn to_range(&self, db: &(dyn ExpandDatabase + 'static)) -> TextRange

pub fn to_in_file_node( &self, db: &(dyn ExpandDatabase + 'static), ) -> InFileWrapper<HirFileId, N>

pub fn to_ptr(&self, db: &(dyn ExpandDatabase + 'static)) -> AstPtr<N>

pub fn erase(&self) -> InFileWrapper<HirFileId, ErasedFileAstId>

pub fn upcast<M>(self) -> InFileWrapper<HirFileId, FileAstId<M>>
where M: AstIdNode, N: Into<M>,

impl InFileWrapper<HirFileId, ErasedFileAstId>

pub fn to_range(&self, db: &(dyn ExpandDatabase + 'static)) -> TextRange

pub fn to_ptr( &self, db: &(dyn ExpandDatabase + 'static), ) -> SyntaxNodePtr<RustLanguage>

impl<FileKind, T> InFileWrapper<FileKind, T>

pub fn new(file_id: FileKind, value: T) -> InFileWrapper<FileKind, T>

pub fn map<F, U>(self, f: F) -> InFileWrapper<FileKind, U>
where F: FnOnce(T) -> U,

impl<FileKind, T> InFileWrapper<FileKind, T>
where FileKind: Copy,

pub fn with_value<U>(&self, value: U) -> InFileWrapper<FileKind, U>

pub fn as_ref(&self) -> InFileWrapper<FileKind, &T>

pub fn borrow<U>(&self) -> InFileWrapper<FileKind, &U>
where T: Borrow<U>,

impl<FileKind, T> InFileWrapper<FileKind, &T>
where FileKind: Copy, T: Clone,

pub fn cloned(&self) -> InFileWrapper<FileKind, T>

impl<FileKind, T> InFileWrapper<FileKind, Option<T>>

pub fn transpose(self) -> Option<InFileWrapper<FileKind, T>>

impl<FileKind, L, R> InFileWrapper<FileKind, Either<L, R>>

pub fn transpose( self, ) -> Either<InFileWrapper<FileKind, L>, InFileWrapper<FileKind, R>>

impl<FileId, T> InFileWrapper<FileId, T>
where FileId: FileIdToSyntax,

pub fn file_syntax( &self, db: &(dyn ExpandDatabase + 'static), ) -> SyntaxNode<RustLanguage>

impl<FileId, N> InFileWrapper<FileId, AstPtr<N>>
where FileId: FileIdToSyntax, N: AstNode,

pub fn to_node(&self, db: &(dyn ExpandDatabase + 'static)) -> N

impl<FileId, N> InFileWrapper<FileId, N>
where FileId: Copy, N: AstNode,

pub fn syntax(&self) -> InFileWrapper<FileId, &SyntaxNode<RustLanguage>>

pub fn node_file_range(&self) -> FileRangeWrapper<FileId>

impl<FileId, N> InFileWrapper<FileId, &N>
where FileId: Copy, N: AstNode,

pub fn syntax_ref(&self) -> InFileWrapper<FileId, &SyntaxNode<RustLanguage>>

impl<FileId, SN> InFileWrapper<FileId, SN>
where FileId: Copy, SN: Borrow<SyntaxNode<RustLanguage>>,

pub fn file_range(&self) -> FileRangeWrapper<FileId>

impl<SN> InFileWrapper<HirFileId, SN>
where SN: Borrow<SyntaxNode<RustLanguage>>,

pub fn parent_ancestors_with_macros( self, db: &(dyn ExpandDatabase + 'static), ) -> impl Iterator<Item = InFileWrapper<HirFileId, SyntaxNode<RustLanguage>>>

pub fn ancestors_with_macros( self, db: &(dyn ExpandDatabase + 'static), ) -> impl Iterator<Item = InFileWrapper<HirFileId, SyntaxNode<RustLanguage>>>

pub fn kind(&self) -> SyntaxKind

pub fn text_range(&self) -> TextRange

pub fn original_file_range_rooted( self, db: &(dyn ExpandDatabase + 'static), ) -> FileRangeWrapper<EditionedFileId>

Falls back to the macro call range if the node cannot be mapped up fully.

For attributes and derives, this will point back to the attribute only. For the entire item use InFile::original_file_range_full.

pub fn original_file_range_with_macro_call_input( self, db: &(dyn ExpandDatabase + 'static), ) -> FileRangeWrapper<EditionedFileId>

Falls back to the macro call range if the node cannot be mapped up fully.

pub fn original_syntax_node_rooted( self, db: &(dyn ExpandDatabase + 'static), ) -> Option<InFileWrapper<EditionedFileId, SyntaxNode<RustLanguage>>>

impl InFileWrapper<HirFileId, &SyntaxNode<RustLanguage>>

pub fn original_file_range_opt( self, db: &(dyn ExpandDatabase + 'static), ) -> Option<(FileRangeWrapper<EditionedFileId>, SyntaxContext)>

Attempts to map the syntax node back up its macro calls.

impl InFileWrapper<MacroCallId, SyntaxToken<RustLanguage>>

pub fn upmap_once( self, db: &(dyn ExpandDatabase + 'static), ) -> InFileWrapper<HirFileId, SmallVec<[TextRange; 1]>>

impl InFileWrapper<HirFileId, SyntaxToken<RustLanguage>>

pub fn original_file_range( self, db: &(dyn ExpandDatabase + 'static), ) -> FileRangeWrapper<EditionedFileId>

Falls back to the macro call range if the node cannot be mapped up fully.

pub fn original_file_range_opt( self, db: &(dyn ExpandDatabase + 'static), ) -> Option<FileRangeWrapper<EditionedFileId>>

Attempts to map the syntax node back up its macro calls.

impl InFileWrapper<MacroCallId, TextSize>

pub fn original_file_range( self, db: &(dyn ExpandDatabase + 'static), ) -> (FileRangeWrapper<EditionedFileId>, SyntaxContext)

impl InFileWrapper<HirFileId, TextRange>

pub fn original_node_file_range( self, db: &(dyn ExpandDatabase + 'static), ) -> (FileRangeWrapper<EditionedFileId>, SyntaxContext)

pub fn original_node_file_range_rooted( self, db: &(dyn ExpandDatabase + 'static), ) -> FileRangeWrapper<EditionedFileId>

pub fn original_node_file_range_with_macro_call_input( self, db: &(dyn ExpandDatabase + 'static), ) -> FileRangeWrapper<EditionedFileId>

pub fn original_node_file_range_opt( self, db: &(dyn ExpandDatabase + 'static), ) -> Option<(FileRangeWrapper<EditionedFileId>, SyntaxContext)>

pub fn original_node_file_range_rooted_opt( self, db: &(dyn ExpandDatabase + 'static), ) -> Option<FileRangeWrapper<EditionedFileId>>

impl<N> InFileWrapper<HirFileId, N>
where N: AstNode,

pub fn original_ast_node_rooted( self, db: &(dyn ExpandDatabase + 'static), ) -> Option<InFileWrapper<EditionedFileId, N>>

impl<T> InFileWrapper<HirFileId, T>

pub fn into_real_file( self, ) -> Result<InFileWrapper<EditionedFileId, T>, InFileWrapper<HirFileId, T>>