Struct hir::Module

pub struct Module { /* private fields */ }

Implementations

impl Module

NB: Module is !HasSource, because it has two source nodes at the same time: definition and declaration.

pub fn definition_source(self, db: &dyn HirDatabase) -> InFile<ModuleSource>

Returns a node which defines this module. That is, a file or a mod foo {} with items.

pub fn is_mod_rs(self, db: &dyn HirDatabase) -> bool

pub fn is_inline(self, db: &dyn HirDatabase) -> bool

pub fn declaration_source(self, db: &dyn HirDatabase) -> Option<InFile<Module>>

Returns a node which declares this module, either a mod foo; or a mod foo {}. None for the crate root.

impl Module

pub fn name(self, db: &dyn HirDatabase) -> Option<Name>

Name of this module.

pub fn krate(self) -> Crate

Returns the crate this module is part of.

pub fn crate_root(self, db: &dyn HirDatabase) -> Module

Topmost parent of this module. Every module has a crate_root, but some might be missing krate. This can happen if a module’s file is not included in the module tree of any target in Cargo.toml.

pub fn is_crate_root(self, db: &dyn HirDatabase) -> bool

pub fn children(self, db: &dyn HirDatabase) -> impl Iterator<Item = Module>

Iterates over all child modules.

pub fn parent(self, db: &dyn HirDatabase) -> Option<Module>

Finds a parent module.

pub fn nearest_non_block_module(self, db: &dyn HirDatabase) -> Module

Finds nearest non-block ancestor Module (self included).

pub fn path_to_root(self, db: &dyn HirDatabase) -> Vec<Module>

pub fn scope( self, db: &dyn HirDatabase, visible_from: Option<Module> ) -> Vec<(Name, ScopeDef)>

Returns a ModuleScope: a set of items, visible in this module.

pub fn diagnostics(self, db: &dyn HirDatabase, acc: &mut Vec<AnyDiagnostic>)

Fills acc with the module’s diagnostics.

pub fn declarations(self, db: &dyn HirDatabase) -> Vec<ModuleDef>

pub fn legacy_macros(self, db: &dyn HirDatabase) -> Vec<Macro>

pub fn impl_defs(self, db: &dyn HirDatabase) -> Vec<Impl>

pub fn find_use_path( self, db: &dyn DefDatabase, item: impl Into<ItemInNs>, prefer_no_std: bool ) -> Option<ModPath>

Finds a path that can be used to refer to the given item from within this module, if possible.

pub fn find_use_path_prefixed( self, db: &dyn DefDatabase, item: impl Into<ItemInNs>, prefix_kind: PrefixKind, prefer_no_std: bool ) -> Option<ModPath>

Finds a path that can be used to refer to the given item from within this module, if possible. This is used for returning import paths for use-statements.