Struct hir::SemanticsImpl

pub struct SemanticsImpl<'db> {
    pub db: &'db dyn HirDatabase,
    /* private fields */
}

Fields
db: &'db dyn HirDatabase
Implementations

impl<'db> SemanticsImpl<'db>

pub fn parse(&self, file_id: FileId) -> SourceFile

pub fn parse_or_expand(&self, file_id: HirFileId) -> SyntaxNode

pub fn expand(&self, macro_call: &MacroCall) -> Option<SyntaxNode>

pub fn expand_attr_macro(&self, item: &Item) -> Option<SyntaxNode>

If item has an attribute macro attached to it, expands it.

pub fn expand_derive_as_pseudo_attr_macro( &self, attr: &Attr ) -> Option<SyntaxNode>

pub fn resolve_derive_macro(&self, attr: &Attr) -> Option<Vec<Option<Macro>>>

pub fn expand_derive_macro(&self, attr: &Attr) -> Option<Vec<SyntaxNode>>

pub fn is_derive_annotated(&self, adt: &Adt) -> bool

pub fn is_attr_macro_call(&self, item: &Item) -> bool

pub fn speculative_expand( &self, actual_macro_call: &MacroCall, speculative_args: &TokenTree, token_to_map: SyntaxToken ) -> Option<(SyntaxNode, SyntaxToken)>

Expand the macro call with a different token tree, mapping the token_to_map down into the expansion. token_to_map should be a token from the speculative args node.

pub fn speculative_expand_attr_macro( &self, actual_macro_call: &Item, speculative_args: &Item, token_to_map: SyntaxToken ) -> Option<(SyntaxNode, SyntaxToken)>

Expand the macro call with a different item as the input, mapping the token_to_map down into the expansion. token_to_map should be a token from the speculative args node.

pub fn speculative_expand_derive_as_pseudo_attr_macro( &self, actual_macro_call: &Attr, speculative_args: &Attr, token_to_map: SyntaxToken ) -> Option<(SyntaxNode, SyntaxToken)>

pub fn as_format_args_parts( &self, string: &String ) -> Option<Vec<(TextRange, Option<PathResolution>)>>

pub fn check_for_format_args_template( &self, original_token: SyntaxToken, offset: TextSize ) -> Option<(TextRange, Option<PathResolution>)>

pub fn descend_node_into_attributes<N: AstNode>( &self, node: N ) -> SmallVec<[N; 1]>

Maps a node down by mapping its first and last token down.

pub fn descend_into_macros( &self, mode: DescendPreference, token: SyntaxToken ) -> SmallVec<[SyntaxToken; 1]>

Descend the token into its macro call if it is part of one, returning the tokens in the expansion that it is associated with.

pub fn descend_into_macros_single( &self, mode: DescendPreference, token: SyntaxToken ) -> SyntaxToken

pub fn original_range(&self, node: &SyntaxNode) -> FileRange

Attempts to map the node out of macro expanded files returning the original file range. If upmapping is not possible, this will fall back to the range of the macro call of the macro file the node resides in.

pub fn original_range_opt(&self, node: &SyntaxNode) -> Option<FileRange>

Attempts to map the node out of macro expanded files returning the original file range.

pub fn original_ast_node<N: AstNode>(&self, node: N) -> Option<N>

Attempts to map the node out of macro expanded files. This only work for attribute expansions, as other ones do not have nodes as input.

pub fn original_syntax_node_rooted( &self, node: &SyntaxNode ) -> Option<SyntaxNode>

Attempts to map the node out of macro expanded files. This only work for attribute expansions, as other ones do not have nodes as input.

pub fn diagnostics_display_range(&self, src: InFile<SyntaxNodePtr>) -> FileRange

pub fn ancestors_with_macros( &self, node: SyntaxNode ) -> impl Iterator<Item = SyntaxNode> + Clone + '_

Iterates the ancestors of the given node, climbing up macro expansions while doing so.

pub fn ancestors_at_offset_with_macros( &self, node: &SyntaxNode, offset: TextSize ) -> impl Iterator<Item = SyntaxNode> + '_

pub fn resolve_lifetime_param( &self, lifetime: &Lifetime ) -> Option<LifetimeParam>

pub fn resolve_label(&self, lifetime: &Lifetime) -> Option<Label>

pub fn resolve_type(&self, ty: &Type) -> Option<Type>

pub fn resolve_trait(&self, path: &Path) -> Option<Trait>

pub fn expr_adjustments(&self, expr: &Expr) -> Option<Vec<Adjustment>>

pub fn type_of_expr(&self, expr: &Expr) -> Option<TypeInfo>

pub fn type_of_pat(&self, pat: &Pat) -> Option<TypeInfo>

pub fn type_of_binding_in_pat(&self, pat: &IdentPat) -> Option<Type>

It also includes the changes that binding mode makes in the type. For example in let ref x @ Some(_) = None the result of type_of_pat is Option<T> but the result of this function is &mut Option<T>

pub fn type_of_self(&self, param: &SelfParam) -> Option<Type>

pub fn pattern_adjustments(&self, pat: &Pat) -> SmallVec<[Type; 1]>

pub fn binding_mode_of_pat(&self, pat: &IdentPat) -> Option<BindingMode>

pub fn resolve_expr_as_callable(&self, call: &Expr) -> Option<Callable>

pub fn resolve_method_call(&self, call: &MethodCallExpr) -> Option<Function>

pub fn resolve_method_call_fallback( &self, call: &MethodCallExpr ) -> Option<Either<Function, Field>>

Attempts to resolve this call expression as a method call falling back to resolving it as a field.

pub fn resolve_method_call_as_callable( &self, call: &MethodCallExpr ) -> Option<Callable>

pub fn resolve_field( &self, field: &FieldExpr ) -> Option<Either<Field, TupleField>>

pub fn resolve_field_fallback( &self, field: &FieldExpr ) -> Option<Either<Either<Field, TupleField>, Function>>

pub fn resolve_record_field( &self, field: &RecordExprField ) -> Option<(Field, Option<Local>, Type)>

pub fn resolve_record_pat_field( &self, field: &RecordPatField ) -> Option<(Field, Type)>

pub fn resolve_macro_call(&self, macro_call: &MacroCall) -> Option<Macro>

pub fn is_proc_macro_call(&self, macro_call: &MacroCall) -> bool

pub fn is_unsafe_macro_call(&self, macro_call: &MacroCall) -> bool

pub fn resolve_attr_macro_call(&self, item: &Item) -> Option<Macro>

pub fn resolve_path(&self, path: &Path) -> Option<PathResolution>

pub fn resolve_bind_pat_to_const(&self, pat: &IdentPat) -> Option<ModuleDef>

pub fn record_literal_missing_fields( &self, literal: &RecordExpr ) -> Vec<(Field, Type)>

pub fn record_pattern_missing_fields( &self, pattern: &RecordPat ) -> Vec<(Field, Type)>

pub fn to_def<T: ToDef>(&self, src: &T) -> Option<T::Def>

pub fn scope(&self, node: &SyntaxNode) -> Option<SemanticsScope<'db>>

pub fn scope_at_offset( &self, node: &SyntaxNode, offset: TextSize ) -> Option<SemanticsScope<'db>>

pub fn source<Def: HasSource>(&self, def: Def) -> Option<InFile<Def::Ast>>
where Def::Ast: AstNode,

Search for a definition’s source and cache its syntax tree

pub fn assert_contains_node(&self, node: &SyntaxNode)

pub fn is_unsafe_method_call(&self, method_call_expr: &MethodCallExpr) -> bool

pub fn is_unsafe_ref_expr(&self, ref_expr: &RefExpr) -> bool

pub fn is_unsafe_ident_pat(&self, ident_pat: &IdentPat) -> bool

pub fn is_inside_unsafe(&self, expr: &Expr) -> bool

Returns true if the node is inside an unsafe context.