hir
Struct SemanticsScope

pub struct SemanticsScope<'a> {
    pub db: &'a dyn HirDatabase,
    /* private fields */
}

SemanticsScope encapsulates the notion of a scope (the set of visible names) at a particular program point.

It is a bit tricky, as scopes do not really exist inside the compiler. Rather, the compiler directly computes for each reference the definition it refers to. It might transiently compute the explicit scope map while doing so, but, generally, this is not something left after the analysis.

However, we do very much need explicit scopes for IDE purposes – completion, at its core, lists the contents of the current scope. The notion of scope is also useful to answer questions like “what would be the meaning of this piece of code if we inserted it into this position?”.

So SemanticsScope is constructed from a specific program point (a syntax node or just a raw offset) and provides access to the set of visible names on a somewhat best-effort basis.

Note that if you are wondering “what does this specific existing name mean?”, you’d better use the resolve_ family of methods.

Fields
db: &'a dyn HirDatabase
Implementations

impl SemanticsScope<'_>

pub fn module(&self) -> Module

pub fn krate(&self) -> Crate

pub fn visible_traits(&self) -> VisibleTraits

Note: VisibleTraits should be treated as an opaque type, passed into `Type

pub fn process_all_names(&self, f: &mut dyn FnMut(Name, ScopeDef))

Calls the passed closure f on all names in scope.

pub fn speculative_resolve(&self, ast_path: &Path) -> Option<PathResolution>

Resolve a path as-if it was written at the given scope. This is necessary a heuristic, as it doesn’t take hygiene into account.

pub fn resolve_mod_path(&self, path: &ModPath) -> impl Iterator<Item = ItemInNs>

pub fn assoc_type_shorthand_candidates<R>( &self, resolution: &PathResolution, cb: impl FnMut(&Name, TypeAlias) -> Option<R>, ) -> Option<R>

Iterates over associated types that may be specified after the given path (using Ty::Assoc syntax).

pub fn generic_def(&self) -> Option<GenericDef>

pub fn extern_crates(&self) -> impl Iterator<Item = (Name, Module)> + '_

pub fn extern_crate_decls(&self) -> impl Iterator<Item = Name> + '_

pub fn has_same_self_type(&self, other: &SemanticsScope<'_>) -> bool