Struct hir::TypeParam

pub struct TypeParam { /* private fields */ }

Implementations

impl TypeParam

pub fn merge(self) -> TypeOrConstParam

pub fn name(self, db: &dyn HirDatabase) -> Name

pub fn module(self, db: &dyn HirDatabase) -> Module

pub fn is_implicit(self, db: &dyn HirDatabase) -> bool

Is this type parameter implicitly introduced (eg. Self in a trait or an impl Trait argument)?

pub fn ty(self, db: &dyn HirDatabase) -> Type

pub fn trait_bounds(self, db: &dyn HirDatabase) -> Vec<Trait>

FIXME: this only lists trait bounds from the item defining the type parameter, not additional bounds that might be added e.g. by a method if the parameter comes from an impl!

pub fn default(self, db: &dyn HirDatabase) -> Option<Type>