hir

Module term_search

Term search

Modules

expr 🔒

Type tree for term search

tactics 🔒

Tactics for term search

Structs

LookupTable 🔒

Lookup table for term search

TermSearchConfig

Configuration options for the term search

TermSearchCtx

Context for the term_search function

Enums

AlternativeExprs 🔒

Helper enum to squash big number of alternative trees into Many variant as there is too many to take into account.

Expr

Type tree shows how can we get from set of types to some type.

NewTypesKey 🔒

Key for lookup table to query new types reached.

Functions

term_search

Term search