hir::term_search
Enum Expr

pub enum Expr {

    Const(Const),
    Static(Static),
    Local(Local),
    ConstParam(ConstParam),
    FamousType {
        ty: Type,
        value: &'static str,
    },
    Function {
        func: Function,
        generics: Vec<Type>,
        params: Vec<Expr>,
    },
    Method {
        func: Function,
        generics: Vec<Type>,
        target: Box<Expr>,
        params: Vec<Expr>,
    },
    Variant {
        variant: Variant,
        generics: Vec<Type>,
        params: Vec<Expr>,
    },
    Struct {
        strukt: Struct,
        generics: Vec<Type>,
        params: Vec<Expr>,
    },
    Tuple {
        ty: Type,
        params: Vec<Expr>,
    },
    Field {
        expr: Box<Expr>,
        field: Field,
    },
    Reference(Box<Expr>),
    Many(Type),

}

Type tree shows how can we get from set of types to some type.

Consider the following code as an example

fn foo(x: i32, y: bool) -> Option<i32> { None }
fn bar() {
   let a = 1;
   let b = true;
   let c: Option<i32> = _;
}

If we generate type tree in the place of _ we get

      Option<i32>
          |
    foo(i32, bool)
     /        \
 a: i32      b: bool

So in short it pretty much gives us a way to get type Option<i32> using the items we have in scope.

Variants

Const(Const)

Constant

Static(Static)

Static variable

Local(Local)

Local variable

ConstParam(ConstParam)

Constant generic parameter

FamousType

Well known type (such as true for bool)

Fields

ty: Type

value: &'static str

Function

Function call (does not take self param)

Fields

func: Function

generics: Vec<Type>

params: Vec<Expr>

Method

Method call (has self param)

Fields

func: Function

generics: Vec<Type>

target: Box<Expr>

params: Vec<Expr>

Variant

Enum variant construction

Fields

variant: Variant

generics: Vec<Type>

params: Vec<Expr>

Struct

Struct construction

Fields

strukt: Struct

generics: Vec<Type>

params: Vec<Expr>

Tuple

Tuple construction

Fields

ty: Type

params: Vec<Expr>

Field

Struct field access

Fields

expr: Box<Expr>

field: Field

Reference(Box<Expr>)

Passing type as reference (with &)

Many(Type)

Indicates possibility of many different options that all evaluate to ty

Implementations

impl Expr

pub fn gen_source_code( &self, sema_scope: &SemanticsScope<'_>, many_formatter: &mut dyn FnMut(&Type) -> String, cfg: ImportPathConfig, edition: Edition, ) -> Result<String, DisplaySourceCodeError>

Generate source code for type tree.

Note that trait imports are not added to generated code. To make sure that the code is valid, callee has to also ensure that all the traits listed by traits_used method are also imported.

pub fn ty(&self, db: &dyn HirDatabase) -> Type

Get type of the type tree.

Same as getting the type of root node

pub fn traits_used(&self, db: &dyn HirDatabase) -> Vec<Trait>

List the traits used in type tree

pub fn is_many(&self) -> bool

Helper function to check if outermost type tree is Expr::Many variant