hir::term_search

Struct LookupTable

struct LookupTable<'db> {
    data: FxHashMap<Type<'db>, AlternativeExprs<'db>>,
    new_types: FxHashMap<NewTypesKey, Vec<Type<'db>>>,
    types_wishlist: FxHashSet<Type<'db>>,
    many_threshold: usize,
}

Lookup table for term search

Lookup table keeps all the state during term search. This means it knows what types and how are reachable.

The secondary functionality for lookup table is to keep track of new types reached since last iteration as well as keeping track of which ScopeDef items have been used. Both of them are to speed up the term search by leaving out types / ScopeDefs that likely do not produce any new results.

Fields
data: FxHashMap<Type<'db>, AlternativeExprs<'db>>

All the Exprs in “value” produce the type of “key”

new_types: FxHashMap<NewTypesKey, Vec<Type<'db>>>

New types reached since last query by the NewTypesKey

types_wishlist: FxHashSet<Type<'db>>

Types queried but not present

many_threshold: usize

Threshold to squash trees to Many

Implementations

impl<'db> LookupTable<'db>

fn new(many_threshold: usize, goal: Type<'db>) -> Self

Initialize lookup table

fn find( &mut self, db: &'db dyn HirDatabase, ty: &Type<'db>, ) -> Option<Vec<Expr<'db>>>

Find all Exprs that unify with the ty

fn find_autoref( &mut self, db: &'db dyn HirDatabase, ty: &Type<'db>, ) -> Option<Vec<Expr<'db>>>

Same as find but automatically creates shared reference of types in the lookup

For example if we have type i32 in data and we query for &i32 it map all the type trees we have for i32 with Expr::Reference and returns them.

fn insert(&mut self, ty: Type<'db>, exprs: impl Iterator<Item = Expr<'db>>)

Insert new type trees for type

Note that the types have to be the same, unification is not enough as unification is not transitive. For example Vec<i32> and FxHashSet<i32> both unify with Iterator<Item = i32>, but they clearly do not unify themselves.

fn iter_types(&self) -> impl Iterator<Item = Type<'db>> + '_

Iterate all the reachable types

fn new_types(&mut self, key: NewTypesKey) -> Vec<Type<'db>>

Query new types reached since last query by key

Create new key if you wish to query it to avoid conflicting with existing queries.

fn types_wishlist(&mut self) -> &FxHashSet<Type<'db>>

Types queried but not found