hir::term_search

Module tactics

Tactics for term search

All the tactics take following arguments

ctx - Context for the term search

defs - Set of items in scope at term search target location

lookup - Lookup table for types

should_continue - Function that indicates when to stop iterating

And they return iterator that yields type trees that unify with the goal type.

Functions

assoc_const 🔒

Associated constant tactic

data_constructor 🔒

Data constructor tactic

famous_types 🔒

Famous types tactic

free_function 🔒

Free function tactic

impl_method 🔒

Impl method tactic

impl_static_method 🔒

Impl static method (without self type) tactic

make_tuple 🔒

Make tuple tactic

struct_projection 🔒

Struct projection tactic

trivial 🔒

Trivial tactic