Module hir::db

Re-exports various subcrates databases so that the calling code can depend only on hir. This breaks abstraction boundary a bit, it would be cool if we didn’t do that.

But we need this for at least LRU caching at the query level.

Structs

AdtDatumQuery

AdtVarianceQuery

AssociatedTyDataQuery

AssociatedTyValueQuery

AstIdMapQuery

AttrsQuery

BlockDefMapQuery

BlockItemTreeQuery

BodyQuery

BodyWithSourceMapQuery

BorrowckQuery

CallableItemSignatureQuery

ConstDataQuery

ConstEvalDiscriminantQuery

ConstEvalQuery

ConstEvalStaticQuery

ConstParamTyQuery

ConstVisibilityQuery

CrateDefMapQuery

CrateLangItemsQuery

CrateNotableTraitsQuery

CrateSupportsNoStdQuery

DeclMacroExpanderQuery

DefDatabaseStorage

Representative struct for the query group.

EnumDataQuery

EnumVariantDataWithDiagnosticsQuery

ExpandDatabaseStorage

Representative struct for the query group.

ExpandProcMacroQuery

ExprScopesQuery

ExternCrateDeclDataQuery

FieldTypesQuery

FieldVisibilitiesQuery

FieldsAttrsQuery

FieldsAttrsSourceMapQuery

FileItemTreeQuery

FnDefDatumQuery

FnDefVarianceQuery

FunctionDataQuery

FunctionVisibilityQuery

GenericDefaultsQuery

GenericParamsQuery

GenericPredicatesForParamQuery

GenericPredicatesQuery

HirDatabaseStorage

Representative struct for the query group.

ImplDataWithDiagnosticsQuery

ImplDatumQuery

ImplSelfTyQuery

ImplTraitQuery

ImportMapQuery

IncoherentInherentImplCratesQuery

InherentImplsInBlockQuery

InherentImplsInCrateQuery

InternAnonymousConstQuery

InternBlockQuery

InternCallableDefQuery

InternClosureQuery

InternConstQuery

InternCoroutineQuery

InternDatabaseStorage

Representative struct for the query group.

InternEnumQuery

InternExternBlockQuery

InternExternCrateQuery

InternFunctionQuery

InternImplQuery

InternImplTraitIdQuery

InternInTypeConstQuery

InternLifetimeParamIdQuery

InternMacro2Query

InternMacroCallQuery

InternMacroRulesQuery

InternProcMacroQuery

InternStaticQuery

InternStructQuery

InternSyntaxContextQuery

InternTraitAliasQuery

InternTraitQuery

InternTypeAliasQuery

InternTypeOrConstParamIdQuery

InternUnionQuery

InternUseQuery

LangItemQuery

LayoutOfAdtQuery

Macro2DataQuery

MacroArgQuery

MacroRulesDataQuery

MirBodyQuery

ParseMacroExpansionErrorQuery

ParseMacroExpansionQuery

ProcMacroDataQuery

ProcMacrosQuery

ProgramClausesForChalkEnvQuery

RealSpanMapQuery

ReturnTypeImplTraitsQuery

StaticDataQuery

StructDataWithDiagnosticsQuery

TargetDataLayoutQuery

TraitAliasDataQuery

TraitDataWithDiagnosticsQuery

TraitDatumQuery

TraitEnvironmentQuery

TraitImplsInBlockQuery

TraitImplsInCrateQuery

TraitImplsInDepsQuery

TyQuery

TypeAliasDataQuery

UnionDataWithDiagnosticsQuery

ValueTyQuery

Traits

DefDatabase

ExpandDatabase

HirDatabase

InternDatabase