Module hir::db

Re-exports various subcrates databases so that the calling code can depend only on hir. This breaks abstraction boundary a bit, it would be cool if we didn’t do that.

But we need this for at least LRU caching at the query level.

Structs

AdtVarianceQuery

AssociatedTyDataQuery

AssociatedTyValueQuery

AstDatabaseStorage

Representative struct for the query group.

AstIdMapQuery

AttrsQuery

BlockDefMapQuery

BodyQuery

BodyWithSourceMapQuery

CallableItemSignatureQuery

ConstDataQuery

ConstEvalQuery

ConstParamTyQuery

ConstVisibilityQuery

CrateDefMapQueryQuery

CrateLangItemsQuery

CrateLimits

CrateSupportsNoStdQuery

DefDatabaseGroupStorage__

DefDatabaseStorage

Representative struct for the query group.

EnableProcAttrMacrosQuery

EnumDataQuery

ExprScopesQuery

FieldTypesQuery

FieldVisibilitiesQuery

FieldsAttrsQuery

FieldsAttrsSourceMapQuery

FileItemTreeQuery

FnDefDatumQuery

FnDefVarianceQuery

FunctionDataQuery

FunctionVisibilityQuery

GenericDefaultsQuery

GenericParamsQuery

GenericPredicatesForParamQuery

GenericPredicatesQuery

HirDatabaseGroupStorage__

HirDatabaseStorage

Representative struct for the query group.

HygieneFrameQuery

ImplDataQuery

ImplDatumQuery

ImplSelfTyQuery

ImplTraitQuery

ImportMapQuery

InferQueryQuery

InherentImplCratesQuery

InherentImplsInBlockQuery

InherentImplsInCrateQuery

InternBlockLookupQuery

InternBlockQuery

InternCallableDefLookupQuery

InternCallableDefQuery

InternClosureLookupQuery

InternClosureQuery

InternConstLookupQuery

InternConstQuery

InternDatabaseGroupStorage__

InternDatabaseStorage

Representative struct for the query group.

InternEnumLookupQuery

InternEnumQuery

InternExternBlockLookupQuery

InternExternBlockQuery

InternFunctionLookupQuery

InternFunctionQuery

InternImplLookupQuery

InternImplQuery

InternImplTraitIdLookupQuery

InternImplTraitIdQuery

InternLifetimeParamIdLookupQuery

InternLifetimeParamIdQuery

InternMacro2LookupQuery

InternMacro2Query

InternMacroCallQuery

InternMacroRulesLookupQuery

InternMacroRulesQuery

InternProcMacroLookupQuery

InternProcMacroQuery

InternStaticLookupQuery

InternStaticQuery

InternStructLookupQuery

InternStructQuery

InternTraitLookupQuery

InternTraitQuery

InternTypeAliasLookupQuery

InternTypeAliasQuery

InternTypeOrConstParamIdLookupQuery

InternTypeOrConstParamIdQuery

InternUnionLookupQuery

InternUnionQuery

InternedCallableDefId

This exists just for Chalk, because Chalk just has a single FnDefId where we have different IDs for struct and enum variant constructors.

InternedClosureId

InternedConstParamId

InternedLifetimeParamId

InternedOpaqueTyId

InternedTypeOrConstParamId

LangItemQuery

Macro2DataQuery

MacroArgTextQuery

MacroDefQuery

MacroExpandQuery

MacroRulesDataQuery

ParseMacroExpansionQuery

ProcMacroDataQuery

ProgramClausesForChalkEnvQuery

ReturnTypeImplTraitsQuery

StaticDataQuery

StructDataQuery

StructDatumQuery

TraitDataQuery

TraitDatumQuery

TraitEnvironmentQuery

TraitImplsInBlockQuery

TraitImplsInCrateQuery

TraitImplsInDepsQuery

TraitSolveQueryQuery

TyQuery

TypeAliasDataQuery

UnionDataQuery

ValueTyQuery

VariantsAttrsQuery

VariantsAttrsSourceMapQuery

Traits

AstDatabase

DefDatabase

HirDatabase

InternDatabase