hir

Module import_map

A map of all publicly exported items in a crate.

Structs

ImportInfo

Item import details stored in the ImportMap.

ImportMap

A map from publicly exported items to its name.

Query

Enums

AssocSearchMode

Three possible ways to search for the name in associated and/or other items.

SearchMode

A way to match import map contents against the search query.

Functions

search_dependencies

Searches dependencies of krate for an importable name matching query.