hir

Module mir

MIR definitions and implementation

Structs

BasicBlock

BorrowckResult

Evaluator

Local

MirBody

Operand

An operand in MIR represents a “value” in Rust, the definition of which is undecided and part of the memory model. One proposal for a definition of values can be found on UCG.

Place

ProjectionId

ProjectionStore

Statement

SwitchTargets

Terminator

VTableMap

Enums

AggregateKind

BinOp

BorrowKind

CastKind

MirEvalError

MirLowerError

MirSpan

MutBorrowKind

MutabilityReason

Stores spans which implies that the local should be mutable.

OperandKind

ProjectionElem

Rvalue

StatementKind

TerminatorKind

UnOp

Functions

borrowck_query

interpret_mir

lower_to_mir

mir_body_for_closure_query

mir_body_query

monomorphized_mir_body_for_closure_query

monomorphized_mir_body_query

pad16

render_const_using_debug_impl

Type Aliases

BasicBlockId

LocalId