hir

Struct Semantics

pub struct Semantics<'db, DB: ?Sized> {
    pub db: &'db DB,
    /* private fields */
}

Primary API to get semantic information, like types, from syntax trees.

Fields
db: &'db DB
Implementations

impl Semantics<'_, dyn HirDatabase>

pub fn new_dyn(db: &dyn HirDatabase) -> Semantics<'_, dyn HirDatabase>

Creates an instance that’s weakly coupled to its underlying database type.

impl<DB: HirDatabase> Semantics<'_, DB>

pub fn new(db: &DB) -> Semantics<'_, DB>

Creates an instance that’s strongly coupled to its underlying database type.

impl<DB: HirDatabase + ?Sized> Semantics<'_, DB>

pub fn hir_file_for(&self, syntax_node: &SyntaxNode) -> HirFileId

pub fn token_ancestors_with_macros( &self, token: SyntaxToken, ) -> impl Iterator<Item = SyntaxNode> + '_

pub fn find_node_at_offset_with_macros<N: AstNode>( &self, node: &SyntaxNode, offset: TextSize, ) -> Option<N>

Find an AstNode by offset inside SyntaxNode, if it is inside Macrofile, search up until it is of the target AstNode type

pub fn find_node_at_offset_with_descend<N: AstNode>( &self, node: &SyntaxNode, offset: TextSize, ) -> Option<N>

Find an AstNode by offset inside SyntaxNode, if it is inside MacroCall, descend it and find again

pub fn find_nodes_at_offset_with_descend<'slf, N: AstNode + 'slf>( &'slf self, node: &SyntaxNode, offset: TextSize, ) -> impl Iterator<Item = N> + 'slf

Find an AstNode by offset inside SyntaxNode, if it is inside an attribute macro call, descend it and find again

pub fn find_namelike_at_offset_with_descend<'slf>( &'slf self, node: &SyntaxNode, offset: TextSize, ) -> impl Iterator<Item = NameLike> + 'slf

pub fn resolve_range_pat(&self, range_pat: &RangePat) -> Option<Struct>

pub fn resolve_range_expr(&self, range_expr: &RangeExpr) -> Option<Struct>

pub fn resolve_await_to_poll(&self, await_expr: &AwaitExpr) -> Option<Function>

pub fn resolve_prefix_expr(&self, prefix_expr: &PrefixExpr) -> Option<Function>

pub fn resolve_index_expr(&self, index_expr: &IndexExpr) -> Option<Function>

pub fn resolve_bin_expr(&self, bin_expr: &BinExpr) -> Option<Function>

pub fn resolve_try_expr(&self, try_expr: &TryExpr) -> Option<Function>

pub fn resolve_variant(&self, record_lit: RecordExpr) -> Option<VariantDef>

pub fn file_to_module_def(&self, file: impl Into<FileId>) -> Option<Module>

pub fn file_to_module_defs( &self, file: impl Into<FileId>, ) -> impl Iterator<Item = Module>

pub fn hir_file_to_module_def( &self, file: impl Into<HirFileId>, ) -> Option<Module>

pub fn hir_file_to_module_defs( &self, file: impl Into<HirFileId>, ) -> impl Iterator<Item = Module>

pub fn is_nightly(&self, krate: Crate) -> bool

pub fn to_adt_def(&self, a: &Adt) -> Option<Adt>

pub fn to_const_def(&self, c: &Const) -> Option<Const>

pub fn to_enum_def(&self, e: &Enum) -> Option<Enum>

pub fn to_enum_variant_def(&self, v: &Variant) -> Option<Variant>

pub fn to_fn_def(&self, f: &Fn) -> Option<Function>

pub fn to_impl_def(&self, i: &Impl) -> Option<Impl>

pub fn to_macro_def(&self, m: &Macro) -> Option<Macro>

pub fn to_module_def(&self, m: &Module) -> Option<Module>

pub fn to_static_def(&self, s: &Static) -> Option<Static>

pub fn to_struct_def(&self, s: &Struct) -> Option<Struct>

pub fn to_trait_def(&self, t: &Trait) -> Option<Trait>

pub fn to_type_alias_def(&self, t: &TypeAlias) -> Option<TypeAlias>

pub fn to_union_def(&self, u: &Union) -> Option<Union>