Struct hir::Semantics

pub struct Semantics<'db, DB> {
    pub db: &'db DB,
    // some fields omitted
}

Primary API to get semantic information, like types, from syntax trees.

Fields
db: &'db DB
Implementations

impl<'db, DB: HirDatabase> Semantics<'db, DB>

pub fn new(db: &DB) -> Semantics<'_, DB>

pub fn parse(&self, file_id: FileId) -> SourceFile

pub fn expand(&self, macro_call: &MacroCall) -> Option<SyntaxNode>

pub fn speculative_expand(
&self,
actual_macro_call: &MacroCall,
hypothetical_args: &TokenTree,
token_to_map: SyntaxToken
) -> Option<(SyntaxNode, SyntaxToken)>

pub fn descend_into_macros(&self, token: SyntaxToken) -> SyntaxToken

pub fn descend_node_at_offset<N: AstNode>(
&self,
node: &SyntaxNode,
offset: TextSize
) -> Option<N>

pub fn original_range(&self, node: &SyntaxNode) -> FileRange

pub fn diagnostics_display_range(
&self,
diagnostics: &dyn Diagnostic
) -> FileRange

pub fn ancestors_with_macros(
&self,
node: SyntaxNode
) -> impl Iterator<Item = SyntaxNode> + '_

pub fn ancestors_at_offset_with_macros(
&self,
node: &SyntaxNode,
offset: TextSize
) -> impl Iterator<Item = SyntaxNode> + '_

pub fn find_node_at_offset_with_macros<N: AstNode>(
&self,
node: &SyntaxNode,
offset: TextSize
) -> Option<N>

Find a AstNode by offset inside SyntaxNode, if it is inside Macrofile, search up until it is of the target AstNode type

pub fn find_node_at_offset_with_descend<N: AstNode>(
&self,
node: &SyntaxNode,
offset: TextSize
) -> Option<N>

Find a AstNode by offset inside SyntaxNode, if it is inside MacroCall, descend it and find again

pub fn resolve_lifetime_param(
&self,
lifetime: &Lifetime
) -> Option<LifetimeParam>

pub fn resolve_label(&self, lifetime: &Lifetime) -> Option<Label>

pub fn type_of_expr(&self, expr: &Expr) -> Option<Type>

pub fn type_of_pat(&self, pat: &Pat) -> Option<Type>

pub fn type_of_self(&self, param: &SelfParam) -> Option<Type>

pub fn resolve_method_call(&self, call: &MethodCallExpr) -> Option<Function>

pub fn resolve_method_call_as_callable(
&self,
call: &MethodCallExpr
) -> Option<Callable>

pub fn resolve_field(&self, field: &FieldExpr) -> Option<Field>

pub fn resolve_record_field(
&self,
field: &RecordExprField
) -> Option<(Field, Option<Local>)>

pub fn resolve_record_pat_field(&self, field: &RecordPatField) -> Option<Field>

pub fn resolve_macro_call(&self, macro_call: &MacroCall) -> Option<MacroDef>

pub fn resolve_path(&self, path: &Path) -> Option<PathResolution>

pub fn resolve_extern_crate(&self, extern_crate: &ExternCrate) -> Option<Crate>

pub fn resolve_variant(&self, record_lit: RecordExpr) -> Option<VariantDef>

pub fn resolve_bind_pat_to_const(&self, pat: &IdentPat) -> Option<ModuleDef>

pub fn record_literal_missing_fields(
&self,
literal: &RecordExpr
) -> Vec<(Field, Type)>

pub fn record_pattern_missing_fields(
&self,
pattern: &RecordPat
) -> Vec<(Field, Type)>

pub fn to_def<T: ToDef>(&self, src: &T) -> Option<T::Def>

pub fn to_module_def(&self, file: FileId) -> Option<Module>

pub fn scope(&self, node: &SyntaxNode) -> SemanticsScope<'db>

pub fn scope_at_offset(
&self,
node: &SyntaxNode,
offset: TextSize
) -> SemanticsScope<'db>

pub fn scope_for_def(&self, def: Trait) -> SemanticsScope<'db>

pub fn assert_contains_node(&self, node: &SyntaxNode)

pub fn is_unsafe_method_call(&self, method_call_expr: &MethodCallExpr) -> bool

pub fn is_unsafe_ref_expr(&self, ref_expr: &RefExpr) -> bool

pub fn is_unsafe_ident_pat(&self, ident_pat: &IdentPat) -> bool