hir

Struct Type

pub struct Type<'db> { /* private fields */ }

Implementations

impl<'db> Type<'db>

pub fn new_slice(ty: Self) -> Self

pub fn new_tuple(krate: Crate, tys: &[Self]) -> Self

pub fn is_unit(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_str(&self) -> bool

pub fn is_never(&self) -> bool

pub fn is_mutable_reference(&self) -> bool

pub fn is_reference(&self) -> bool

pub fn contains_reference(&self, db: &'db dyn HirDatabase) -> bool

pub fn as_reference(&self) -> Option<(Type<'db>, Mutability)>

pub fn add_reference(&self, mutability: Mutability) -> Self

pub fn is_slice(&self) -> bool

pub fn is_usize(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_char(&self) -> bool

pub fn is_int_or_uint(&self) -> bool

pub fn is_scalar(&self) -> bool

pub fn is_tuple(&self) -> bool

pub fn remove_ref(&self) -> Option<Type<'db>>

pub fn as_slice(&self) -> Option<Type<'db>>

pub fn strip_references(&self) -> Self

pub fn strip_reference(&self) -> Self

pub fn is_unknown(&self) -> bool

pub fn into_future_output(&self, db: &'db dyn HirDatabase) -> Option<Type<'db>>

Checks that particular type ty implements std::future::IntoFuture or std::future::Future and returns the Output associated type. This function is used in .await syntax completion.

pub fn future_output(self, db: &'db dyn HirDatabase) -> Option<Type<'db>>

This does not resolve IntoFuture, only Future.

pub fn iterator_item(self, db: &'db dyn HirDatabase) -> Option<Type<'db>>

This does not resolve IntoIterator, only Iterator.

pub fn impls_iterator(self, db: &'db dyn HirDatabase) -> bool

pub fn into_iterator_iter(self, db: &'db dyn HirDatabase) -> Option<Type<'db>>

Resolves the projection <Self as IntoIterator>::IntoIter and returns the resulting type

pub fn impls_fnonce(&self, db: &'db dyn HirDatabase) -> bool

Checks that particular type ty implements std::ops::FnOnce.

This function can be used to check if a particular type is callable, since FnOnce is a supertrait of Fn and FnMut, so all callable types implements at least FnOnce.

pub fn impls_trait( &self, db: &'db dyn HirDatabase, trait_: Trait, args: &[Type<'db>], ) -> bool

pub fn normalize_trait_assoc_type( &self, db: &'db dyn HirDatabase, args: &[Type<'db>], alias: TypeAlias, ) -> Option<Type<'db>>

pub fn is_copy(&self, db: &'db dyn HirDatabase) -> bool

pub fn as_callable(&self, db: &'db dyn HirDatabase) -> Option<Callable<'db>>

pub fn is_closure(&self) -> bool

pub fn as_closure(&self) -> Option<Closure<'db>>

pub fn is_fn(&self) -> bool

pub fn is_array(&self) -> bool

pub fn is_packed(&self, db: &'db dyn HirDatabase) -> bool

pub fn is_raw_ptr(&self) -> bool

pub fn remove_raw_ptr(&self) -> Option<Type<'db>>

pub fn contains_unknown(&self) -> bool

pub fn fields(&self, db: &'db dyn HirDatabase) -> Vec<(Field, Self)>

pub fn tuple_fields(&self, _db: &'db dyn HirDatabase) -> Vec<Self>

pub fn as_array(&self, db: &'db dyn HirDatabase) -> Option<(Self, usize)>

pub fn fingerprint_for_trait_impl(&self) -> Option<SimplifiedType>

pub fn autoderef( &self, db: &'db dyn HirDatabase, ) -> impl Iterator<Item = Type<'db>> + use<'_, 'db>

Returns types that this type dereferences to (including this type itself). The returned iterator won’t yield the same type more than once even if the deref chain contains a cycle.

pub fn iterate_assoc_items<T>( &self, db: &'db dyn HirDatabase, callback: impl FnMut(AssocItem) -> Option<T>, ) -> Option<T>

pub fn type_arguments(&self) -> impl Iterator<Item = Type<'db>> + '_

Iterates its type arguments

It iterates the actual type arguments when concrete types are used and otherwise the generic names. It does not include const arguments.

For code, such as:

struct Foo<T, U>

impl<U> Foo<String, U>

It iterates:

- "String"
- "U"

pub fn type_and_const_arguments<'a>( &'a self, db: &'a dyn HirDatabase, display_target: DisplayTarget, ) -> impl Iterator<Item = SmolStr> + 'a

Iterates its type and const arguments

It iterates the actual type and const arguments when concrete types are used and otherwise the generic names.

For code, such as:

struct Foo<T, const U: usize, const X: usize>

impl<U> Foo<String, U, 12>

It iterates:

- "String"
- "U"
- "12"

pub fn generic_parameters<'a>( &'a self, db: &'a dyn HirDatabase, display_target: DisplayTarget, ) -> impl Iterator<Item = SmolStr> + 'a

Combines lifetime indicators, type and constant parameters into a single Iterator

pub fn iterate_method_candidates_with_traits<T>( &self, db: &'db dyn HirDatabase, scope: &SemanticsScope<'_>, traits_in_scope: &FxHashSet<TraitId>, name: Option<&Name>, callback: impl FnMut(Function) -> Option<T>, ) -> Option<T>

pub fn iterate_method_candidates<T>( &self, db: &'db dyn HirDatabase, scope: &SemanticsScope<'_>, name: Option<&Name>, callback: impl FnMut(Function) -> Option<T>, ) -> Option<T>

pub fn iterate_method_candidates_split_inherent( &self, db: &'db dyn HirDatabase, scope: &SemanticsScope<'_>, traits_in_scope: &FxHashSet<TraitId>, name: Option<&Name>, callback: impl MethodCandidateCallback, )

Allows you to treat inherent and non-inherent methods differently.

Note that inherent methods may actually be trait methods! For example, in dyn Trait, the trait’s methods are considered inherent methods.

pub fn iterate_path_candidates<T>( &self, db: &'db dyn HirDatabase, scope: &SemanticsScope<'_>, traits_in_scope: &FxHashSet<TraitId>, name: Option<&Name>, callback: impl FnMut(AssocItem) -> Option<T>, ) -> Option<T>

pub fn iterate_path_candidates_split_inherent( &self, db: &'db dyn HirDatabase, scope: &SemanticsScope<'_>, traits_in_scope: &FxHashSet<TraitId>, name: Option<&Name>, callback: impl PathCandidateCallback, )

Iterates over inherent methods.

In some circumstances, inherent methods methods may actually be trait methods! For example, when dyn Trait is a receiver, trait’s methods would be considered to be inherent methods.

pub fn as_adt(&self) -> Option<Adt>

pub fn as_builtin(&self) -> Option<BuiltinType>

pub fn as_dyn_trait(&self) -> Option<Trait>

pub fn applicable_inherent_traits( &self, db: &'db dyn HirDatabase, ) -> impl Iterator<Item = Trait>

If a type can be represented as dyn Trait, returns all traits accessible via this type, or an empty iterator otherwise.

pub fn env_traits( &self, db: &'db dyn HirDatabase, ) -> impl Iterator<Item = Trait>

pub fn as_impl_traits( &self, db: &'db dyn HirDatabase, ) -> Option<impl Iterator<Item = Trait>>

pub fn as_associated_type_parent_trait( &self, db: &'db dyn HirDatabase, ) -> Option<Trait>

pub fn walk(&self, db: &'db dyn HirDatabase, callback: impl FnMut(Type<'db>))

Visits every type, including generic arguments, in this type. callback is called with type itself first, and then with its generic arguments.

pub fn could_unify_with( &self, db: &'db dyn HirDatabase, other: &Type<'db>, ) -> bool

Check if type unifies with another type.

Note that we consider placeholder types to unify with everything. For example Option<T> and Option<U> unify although there is unresolved goal T = U.

pub fn could_unify_with_deeply( &self, db: &'db dyn HirDatabase, other: &Type<'db>, ) -> bool

Check if type unifies with another type eagerly making sure there are no unresolved goals.

This means that placeholder types are not considered to unify if there are any bounds set on them. For example Option<T> and Option<U> do not unify as we cannot show that T = U

pub fn could_coerce_to(&self, db: &'db dyn HirDatabase, to: &Type<'db>) -> bool

pub fn as_type_param(&self, _db: &'db dyn HirDatabase) -> Option<TypeParam>

pub fn generic_params( &self, db: &'db dyn HirDatabase, ) -> FxHashSet<GenericParam>

Returns unique GenericParams contained in this type.

pub fn layout(&self, db: &'db dyn HirDatabase) -> Result<Layout, LayoutError>

pub fn drop_glue(&self, db: &'db dyn HirDatabase) -> DropGlue