Struct hir::Type

pub struct Type { /* private fields */ }

Implementations

impl Type

pub fn reference(inner: &Type, m: Mutability) -> Type

pub fn new_slice(ty: Type) -> Type

pub fn is_unit(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_never(&self) -> bool

pub fn is_mutable_reference(&self) -> bool

pub fn is_reference(&self) -> bool

pub fn as_reference(&self) -> Option<(Type, Mutability)>

pub fn is_slice(&self) -> bool

pub fn is_usize(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_char(&self) -> bool

pub fn is_int_or_uint(&self) -> bool

pub fn remove_ref(&self) -> Option<Type>

pub fn as_slice(&self) -> Option<Type>

pub fn strip_references(&self) -> Type

pub fn strip_reference(&self) -> Type

pub fn is_unknown(&self) -> bool

pub fn impls_into_future(&self, db: &dyn HirDatabase) -> bool

Checks that particular type ty implements std::future::IntoFuture or std::future::Future. This function is used in .await syntax completion.

pub fn impls_fnonce(&self, db: &dyn HirDatabase) -> bool

Checks that particular type ty implements std::ops::FnOnce.

This function can be used to check if a particular type is callable, since FnOnce is a supertrait of Fn and FnMut, so all callable types implements at least FnOnce.

pub fn impls_trait( &self, db: &dyn HirDatabase, trait_: Trait, args: &[Type] ) -> bool

pub fn normalize_trait_assoc_type( &self, db: &dyn HirDatabase, args: &[Type], alias: TypeAlias ) -> Option<Type>

pub fn is_copy(&self, db: &dyn HirDatabase) -> bool

pub fn as_callable(&self, db: &dyn HirDatabase) -> Option<Callable>

pub fn is_closure(&self) -> bool

pub fn is_fn(&self) -> bool

pub fn is_array(&self) -> bool

pub fn is_packed(&self, db: &dyn HirDatabase) -> bool

pub fn is_raw_ptr(&self) -> bool

pub fn remove_raw_ptr(&self) -> Option<Type>

pub fn contains_unknown(&self) -> bool

pub fn fields(&self, db: &dyn HirDatabase) -> Vec<(Field, Type)>

pub fn tuple_fields(&self, _db: &dyn HirDatabase) -> Vec<Type>

pub fn as_array(&self, _db: &dyn HirDatabase) -> Option<(Type, usize)>

pub fn autoderef<'a>( &'a self, db: &'a dyn HirDatabase ) -> impl Iterator<Item = Type> + 'a

pub fn iterate_assoc_items<T>( &self, db: &dyn HirDatabase, krate: Crate, callback: impl FnMut(AssocItem) -> Option<T> ) -> Option<T>

pub fn type_arguments(&self) -> impl Iterator<Item = Type> + '_

Iterates its type arguments

It iterates the actual type arguments when concrete types are used and otherwise the generic names. It does not include const arguments.

For code, such as:

struct Foo<T, U>

impl<U> Foo<String, U>

It iterates:

- "String"
- "U"

pub fn type_and_const_arguments<'a>( &'a self, db: &'a dyn HirDatabase ) -> impl Iterator<Item = SmolStr> + 'a

Iterates its type and const arguments

It iterates the actual type and const arguments when concrete types are used and otherwise the generic names.

For code, such as:

struct Foo<T, const U: usize, const X: usize>

impl<U> Foo<String, U, 12>

It iterates:

- "String"
- "U"
- "12"

pub fn generic_parameters<'a>( &'a self, db: &'a dyn HirDatabase ) -> impl Iterator<Item = SmolStr> + 'a

Combines lifetime indicators, type and constant parameters into a single Iterator

pub fn iterate_method_candidates_with_traits<T>( &self, db: &dyn HirDatabase, scope: &SemanticsScope<'_>, traits_in_scope: &FxHashSet<TraitId>, with_local_impls: Option<Module>, name: Option<&Name>, callback: impl FnMut(Function) -> Option<T> ) -> Option<T>

pub fn iterate_method_candidates<T>( &self, db: &dyn HirDatabase, scope: &SemanticsScope<'_>, with_local_impls: Option<Module>, name: Option<&Name>, callback: impl FnMut(Function) -> Option<T> ) -> Option<T>

pub fn iterate_path_candidates<T>( &self, db: &dyn HirDatabase, scope: &SemanticsScope<'_>, traits_in_scope: &FxHashSet<TraitId>, with_local_impls: Option<Module>, name: Option<&Name>, callback: impl FnMut(AssocItem) -> Option<T> ) -> Option<T>

pub fn as_adt(&self) -> Option<Adt>

pub fn as_builtin(&self) -> Option<BuiltinType>

pub fn as_dyn_trait(&self) -> Option<Trait>

pub fn applicable_inherent_traits<'a>( &'a self, db: &'a dyn HirDatabase ) -> impl Iterator<Item = Trait> + 'a

If a type can be represented as dyn Trait, returns all traits accessible via this type, or an empty iterator otherwise.

pub fn env_traits<'a>( &'a self, db: &'a dyn HirDatabase ) -> impl Iterator<Item = Trait> + 'a

pub fn as_impl_traits( &self, db: &dyn HirDatabase ) -> Option<impl Iterator<Item = Trait>>

pub fn as_associated_type_parent_trait( &self, db: &dyn HirDatabase ) -> Option<Trait>

pub fn walk(&self, db: &dyn HirDatabase, cb: impl FnMut(Type))

Visits every type, including generic arguments, in this type. cb is called with type itself first, and then with its generic arguments.

pub fn could_unify_with(&self, db: &dyn HirDatabase, other: &Type) -> bool

pub fn could_coerce_to(&self, db: &dyn HirDatabase, to: &Type) -> bool

pub fn as_type_param(&self, db: &dyn HirDatabase) -> Option<TypeParam>

pub fn generic_params(&self, db: &dyn HirDatabase) -> FxHashSet<GenericParam>

Returns unique GenericParams contained in this type.