Struct hir::Type

pub struct Type { /* fields omitted */ }

Implementations

impl Type

pub fn is_unit(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_mutable_reference(&self) -> bool

pub fn is_usize(&self) -> bool

pub fn remove_ref(&self) -> Option<Type>

pub fn strip_references(&self) -> Type

pub fn is_unknown(&self) -> bool

pub fn impls_future(&self, db: &dyn HirDatabase) -> bool

Checks that particular type ty implements std::future::Future. This function is used in .await syntax completion.

pub fn impls_fnonce(&self, db: &dyn HirDatabase) -> bool

Checks that particular type ty implements std::ops::FnOnce.

This function can be used to check if a particular type is callable, since FnOnce is a supertrait of Fn and FnMut, so all callable types implements at least FnOnce.

pub fn impls_trait(
&self,
db: &dyn HirDatabase,
trait_: Trait,
args: &[Type]
) -> bool

pub fn normalize_trait_assoc_type(
&self,
db: &dyn HirDatabase,
args: &[Type],
alias: TypeAlias
) -> Option<Type>

pub fn is_copy(&self, db: &dyn HirDatabase) -> bool

pub fn as_callable(&self, db: &dyn HirDatabase) -> Option<Callable>

pub fn is_closure(&self) -> bool

pub fn is_fn(&self) -> bool

pub fn is_packed(&self, db: &dyn HirDatabase) -> bool

pub fn is_raw_ptr(&self) -> bool

pub fn contains_unknown(&self) -> bool

pub fn fields(&self, db: &dyn HirDatabase) -> Vec<(Field, Type)>

pub fn tuple_fields(&self, _db: &dyn HirDatabase) -> Vec<Type>

pub fn autoderef<'a>(
&'a self,
db: &'a dyn HirDatabase
) -> impl Iterator<Item = Type> + 'a

pub fn iterate_assoc_items<T>(
self,
db: &dyn HirDatabase,
krate: Crate,
callback: impl FnMut(AssocItem) -> Option<T>
) -> Option<T>

pub fn type_arguments(&self) -> impl Iterator<Item = Type> + '_

pub fn iterate_method_candidates<T>(
&self,
db: &dyn HirDatabase,
krate: Crate,
traits_in_scope: &FxHashSet<TraitId>,
name: Option<&Name>,
callback: impl FnMut(&Ty, Function) -> Option<T>
) -> Option<T>

pub fn iterate_path_candidates<T>(
&self,
db: &dyn HirDatabase,
krate: Crate,
traits_in_scope: &FxHashSet<TraitId>,
name: Option<&Name>,
callback: impl FnMut(&Ty, AssocItem) -> Option<T>
) -> Option<T>

pub fn as_adt(&self) -> Option<Adt>

pub fn as_builtin(&self) -> Option<BuiltinType>

pub fn as_dyn_trait(&self) -> Option<Trait>

pub fn applicable_inherent_traits<'a>(
&'a self,
db: &'a dyn HirDatabase
) -> impl Iterator<Item = Trait> + 'a

If a type can be represented as dyn Trait, returns all traits accessible via this type, or an empty iterator otherwise.

pub fn as_impl_traits(&self, db: &dyn HirDatabase) -> Option<Vec<Trait>>

pub fn as_associated_type_parent_trait(
&self,
db: &dyn HirDatabase
) -> Option<Trait>

pub fn walk(&self, db: &dyn HirDatabase, cb: impl FnMut(Type))

pub fn could_unify_with(&self, other: &Type) -> bool